feat: dashboard 추가,  가입자수 데일리 지표 추가

+++ src/services/user.ts
@@ -5,6 +5,7 @@ import { INotificationModel } from '../models/Notification';
 import { IPostModel } from '../models/Post';
 import config from '../config/index';
 import CustomError from '../CustomError';
+import { SignOutUser } from '../models/SignOutUser';
 
 export class UserService {
   constructor(
@@ -38,10 +39,24 @@ export class UserService {
     return { userRecord, accessToken, refreshToken };
   }
 
+  // 회원 탈퇴
   async deleteUser(id: Types.ObjectId, tokenUserId: Types.ObjectId) {
     if (id.toString() !== tokenUserId.toString())
       throw new CustomError('NotAuthenticatedError', 401, 'User does not match');
-    await this.userModel.findOneAndDelete({ _id: id });
+    const user: IUserDocument | null = await this.userModel.findById(id);
+
+    if (user) {
+      // 탈퇴 유저 이력 생성
+      await SignOutUser.create({
+        idToken: user.idToken,
+        tokenType: user.tokenType,
+        nickName: user.nickName,
+        signInDate: user.createdAt,
+        signOutDate: new Date(),
+        userId: user._id,
+      });
+      await this.userModel.findOneAndDelete({ _id: id });
+    }
   }
 
   // 사용자가 관심 등록한 글 리스트를 조회한다.
@@ -102,5 +117,44 @@ export class UserService {
   async addReadLists(postId: Types.ObjectId, userId: Types.ObjectId) {
     const user = await this.userModel.addReadList(postId, userId);
     return user;
+  }
+
+  // 데일리 액션) 현재 총 회원 수, 오늘 가입자, 오늘 탈퇴자
+  async findDashboardDailyUser() {
+    const totalUser: number = await this.userModel.countDocuments();
+    const today = new Date();
+    today.setHours(0, 0, 0, 0);
+    const signUpCount: number = await this.userModel.countDocuments({ createdAt: { $gte: today } });
+    const signOutCount: number = await SignOutUser.countDocuments({ signOutDate: { $gte: today } });
+
+    return {
+      totalUser,
+      signUpCount,
+      signOutCount,
+    };
+  }
+
+  // 일자별 회원 가입 현황(일자 / 신규 가입자 / 탈퇴자)
+  async findDashboardHistoryUser() {
+    const today = new Date('09/01/2022');
+
+    const userHistory = await this.userModel.aggregate([
+      { $match: { createdAt: { $gte: today } } },
+      { $group: { _id: { $dateToString: { format: '%Y-%m-%d', date: '$createdAt' } }, signIn: { $sum: 1 } } },
+      { $addFields: { signOut: 0 } },
+      {
+        $unionWith: {
+          coll: 'signoutusers',
+          pipeline: [
+            { $match: { signOutDate: { $gte: today } } },
+            { $group: { _id: { $dateToString: { format: '%Y-%m-%d', date: '$signOutDate' } }, signOut: { $sum: 1 } } },
+            { $addFields: { signIn: 0 } },
+          ],
+        },
+      },
+      { $group: { _id: '$_id', signIn: { $sum: '$signIn' }, signOut: { $sum: '$signOut' } } },
+      { $sort: { _id: 1 } },
+    ]);
+    return userHistory;
   }
 }
